Pull out worker interface into its own file
