+++ src/ray/raylet/worker_pool.cc
@@ -33,7 +33,9 @@
 #include "ray/common/ray_config.h"
 #include "ray/common/runtime_env_common.h"
 #include "ray/common/status.h"
+#include "ray/core_worker_rpc_client/core_worker_client_interface.h"
 #include "ray/stats/metric_defs.h"
+#include "ray/util/container_util.h"
 #include "ray/util/logging.h"
 #include "ray/util/network_util.h"
 #include "ray/util/time.h"
